week numbers

+++ src/attendanceautomation/BE/AttendanceStatus.java
@@ -79,7 +79,7 @@ public class AttendanceStatus {
     }
 
     public String getStatus() {
-        switch (status.getValue().intValue()) {
+        switch (status.getValue()) {
             case 0:
                 return "🗙";
             case 1:
